Update which improves the code a lot.
Now the code is much cleaner. You provide your inputs struct, last results struct, some modifiers and without the need to worry about the state of the player after calling the function, it returns a struct of results.
Some work on handling hitting and jumping from the ground mid tick started, but it involves some math and I did not spend much time on that.
The code now runs fine. At least when simulating, it handles randomized 70 ms differences of latency, both sending inputs and receiving results. After that, the behavior is glitchy. If the latency stays constant, glitching should not happen.
Added snapping to grid with configurable size. This was done to make floating point errors during calculations not differ from the server that much.
GC allocation still needs to be looked at. Might have increased a bit.