New update bringing strafing
Now, inside the Move function, all the required parts are called using functions. This will make reading everything easier and easier to separate different parts of the code.
Removed EveryFixedUpdate movement type because it was broken and it would only lead to desynced, jittery movement.
Added strafing. Similar to quake/valve games. There are curves which define the acceleration based on turn angle and speed. Can be disabled by commenting out the function call in the Move function.