Fixed network code, added network data analyzer and added support for trusting client results
Network code has been fixed. Now, prediction errors should not happen as frequently.
Network data analyzer has been added. If debug mode is enabled, and the serializer script is enabled on the player, all controller inputs and results are saved to a file (inside the persistent data path). Later, both client and server files can be compared inside editor. It highlights the ticks where network errors have occured so the errors can be found easily. Note, the current approach required to create a new results and inputs classes which use serializeable vectors and quaternions. So, if the original classes are edited, be sure to change the serializeable ones as well.
Added support for trusting client results. Configureable inside the data file, it allows to set the results differences to check against the results to see if they can be accepted as \"close enough\". Inside the controller script, be sure to uncomment CLIENT_TRUST definition, this is so, because no one wants to waste unnecessary bandwidth with the results sent by the client which will never be used.