An update adding third person movement and sliding of slopes
Third person movement added. This uses a CameraControl script which does a lot of nice things. Some of them are in the code but not used, because the script was basically a copy-paste from my game with some changes. Make sure the script runs after the Controller script because otherwise, it will be jittery. See Script Execution Order.
The actual input gathering code moved out of the Controller script. An interface is used for that. DefaultInputManager is the one used at this point. It is made so that custom input gathering could be easily implemented just like the third-person movement.
Sliding of the steep slopes was added. A slide friction variable can be changed to set how much will it slide off.
Input data was separated from the default data.
Head of the character does not rotate. Simple animations with blendshapes should be used.
Some fixes to make playing on high send rate possible.