Removes network-side GC allocations and fixes some bugs
Network-side GC allocations removed. Now, the player sends the inputs using messages and the server sends results using OnSerialize part. Remote clients read the data from OnDeserialize and it gets processed. This also allows to not send unchanged data (not yet applied to the player inputs) so the bandwidth gets even lower. After applying the same method for inputs, the size of data sent while not moving would be 64 bits to the server and 96 bits from the server (plus all the UNet part).
A small bug fixed when the player could get launched into air if it moved up a steep obstacle.