Added AI inputs override and ability to control the player outside the Controller script
AI Inputs override can be used by setting public variable aiEnabled to true, aiTargetReached to 0 or 1 and aiTarget1, aiTarget2 to the AI position targets. This way the camera X rotation is not used, instead the rotation is calculated based on direction to the target. Note: this has not been extensively tested in multiplayer mode and prediction errors could occur due to the way AI target is set.
SetControl function has been added. This allows to tell the Controller that the player is controlled outside the script (useful for moving the player inside vehicles). This way instead of doing any calculations, the MoveCharacter function only updates the position to the current transform position. This can cause some prediction problems when entering/exiting the mode but that is not a big problem.
Updated third person movement to make it a bit better. Some tweaking is needed.