Fixed the inputs sending
Previously, only one remote client could send inputs and others would just be moved the way that client moved. Then, I disabled that but it still did not work properly, because only the last client would move. Now the OnSendInputs is handled by the game manager and it forwards the message to the correct controller.