Implementation of universal object data recording
Now gameplay recordings are recorded differently. A different data structure is used storing a byte array which gets read from and written to by the interface implementing the Init, Tick and SetData functions. It is possible to derrive from the interface implementation to build functionality on top. It is perfectly shown by the player implementation deriving from the base implementation. Even though character position and rotation could be taken from the controller, the base implementation writes them to the buffer and the player implementation adds the rest on top.
Fixed the timestamp text value not changing while paused.
Made network send rate global, added global network settings and the GameManager has to be added somewhere to make everything work. It still is mostly static, but later more code will need to not be static.
Other minor changes.