Simple extensibility with an example, predicted network variables, recordings version bump
Added an ability to extend the controller using inheritance, the list of overridable methods is available in the abstract class base for the controller.
Added custom synchronized variables that the local player is able to predict and use inside the running of the commands, currently only a couple of types are fully implemented, but the rest can be implemented very easily. 32 or 64 (depends on the defines) such variables are supported per controller.
Recordings version bumped to version 3, due to the recent changes in Inputs struct changes and addition of PredVars.
Example of the controller extending has been added.
Fix client-side prediction being wrong due to too small inputs buffer inside the controller data object.