Interpolation rework, movement code separation, speedhack prevention, fixes and optimizations
Interpolation code remade, allowing to interpolate between 3 records. Also detects teleportation if movement distance is too big per tick. Exposed 2 virtual functions: ProcessInterpolation and ProcessTeleportation. The first has 3 arguments: starting results, ending results and normalized time in between them.
Separated movement code to a separate script MovementController. This allows to change the not so spectacular movement code with a custom solution with ease.
Moved AI code to ExampleCharacter since it is not a mandatory feature in many games, but was not worth having a separate script.
Prevention of speedhacking by implementing a sliding window inside of which the scripts can be executed. The client also tries to adjust to keep in sync with the server if something goes wrong.
Optimized gameplay (and lag compensation history) recording not to have any GC allocations. The same goes for testing CFlags. There still might be some GC allocations inside the movement restoring part but that is an improvement already.
Fixed GameManager spamming errors on script reload.
Overall code cleanup.