Fixed issue with spawning on local host, moved PredVar definitions to a separate file, initial work towards predictable physics simulation.
A small issue fixed which was caused due to tempResults not being initialized properly.
All PredVars were moved to a separate file. More was experimented on with them, trying to make it easier to use but nothing was feasible.
Work started on predictable physics simulation which would allow to predict complex physics objects like wehicles without much of a problem.